Abstract Common ImageOIParam class
